Move all parse tests to parser
- Add test utilities to library, but don't make
  this part of the interface.

Locally, it says a lot of tests are no longer run.
I have to verify that this is indeed the case,
otherwise, something is clearly going wrong.
