+++ Cargo.toml
@@ -12,10 +12,6 @@ repository = "https://github.com/JSAbrahams/mamba"
 keywords = ["python", "transpiler", "compiler"]
 categories = ["parsing"]
 
-[badges]
-appveyor = { repository = "JSAbrahams/mamba", branch = "master", service = "github" }
-travis-ci = { repository = "JSAbrahams/mamba", branch = "master" }
-
 [dependencies]
 tempfile = "3.1.0"
 assert_cmd = "0.10"
