+++ src/lib.rs
@@ -13,3 +13,10 @@ pub mod desugar;
 pub mod parse;
 
 pub mod pipeline;
+
+#[cfg(test)]
+mod test_util {
+    // Manual include, otherwise, we have to make this part of the interface to make use of these
+    // utility functions in tests. We don't want this to be part of the interface.
+    include!("../tests/common.rs");
+}
