+++ src/parse/mod.rs
@@ -97,3 +97,17 @@ fn parse_direct(input: &[Lex]) -> ParseResult<Vec<AST>> {
         _ => Ok(vec![])
     }
 }
+
+#[cfg(test)]
+mod test {
+    use crate::parse::lex::tokenize;
+    use crate::parse::parse;
+    use crate::parse::result::ParseResult;
+    use crate::test_util::resource_content;
+
+    #[test]
+    fn parse_empty_file() -> ParseResult<()> {
+        let source = resource_content(true, &[], "empty_file.mamba");
+        parse(&tokenize(&source).unwrap()).map(|_| ())
+    }
+}
