--- tests/parse/invalid/mod.rs
@@ -1,3 +1,0 @@
-pub mod compound;
-mod control_flow;
-mod definition;
