--- tests/parse/valid/mod.rs
@@ -1,5 +1,0 @@
-pub mod class;
-pub mod collection;
-mod control_flow;
-pub mod error;
-pub mod function;
